added two new columns and restructured some code to be in other places (i.e. audit atributes not being sent in for save)
Added a check to mark as Offer rather than Approved when a DA approves a deal in submitted.

However I did discover through testing that bulk updates are not actually working correctly. I rewrote it to work but it is definitely far from efficient and prone to partial failures so it needs to be re-done eventually by making ActionPricingStrategies() independent of contract/customer.  Right now it seems to be reliant on having the same contractID.  need to investigate.



git-tfs-id: [https://tfs-fm-04.devtools.intel.com/tfs/DefaultCollection_TFS04]$/RDMS/MyDeals/MyDeals_Release/UI/main;C1497942